requirements.txt
Pillow
Flask
Flask-CORS
